Revert "008:Fix bug"
