+++ packages/backend/routes/index.js
@@ -1,21 +1,14 @@
 const express = require("express");
 const router = express.Router()
 
-router.get('/', async (req, res) => {
-  try {
-     const query =  {actor_count : {$gt : 1}}
-     const pageSize = 7;
-     const totalFilms = await Films.countDocuments({})
-     const films = await Films.find(query)
-     .limit(pageSize)
- 
-     res.json(films);
-
-  } catch (error) {
-     console.error(error)
-     res.status(500).send("Server Error")
-  }
+router.get('/', (req, res) => {
+  //res.render('index')
+ // res.send("Hello from the CBF Academy backend!");
+ res.json({
+  "name": "Jasmine",
+  "age": 31
+ })
+//  res.sendFile(path.join(__dirname, "../frontend/public", "index.html"))
 })
-
 
 module.exports = router
